move non-reconcile functions out of cluster-controller
